pyproject.toml
[tool.poetry]
name = "qctrl-api"
version = "0.1.0"
description = "RESTful API for BLACK OPAL to manage quantum controls."
authors = ["lamcw <thomas@lamcw.com>"]
license = "Apache-2.0"

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0"
djangorestframework = "^3.11"
django-environ = "^0.4.5"
psycopg2 = "^2.8"
djangorestframework-csv = "^2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
